+++ app/src/main/cpp/src/VorbisFileEncoder.hpp
@@ -26,7 +26,7 @@ class VorbisFileEncoder final {
     vorbis_dsp_state vd; /* central working state for the packet->PCM decoder */
     vorbis_block     vb; /* local working space for packet->PCM decode */
 
-    std::function<void(long)> _soundEnergyLevelListener;
+    std::function<void(float)> _soundEnergyLevelListener;
 
 public:
 
@@ -39,7 +39,7 @@ public:
 
     bool initialize();
 
-    void setSoundEnergyLevelListener(std::function<void(long)> listener);
+    void setSoundEnergyLevelListener(std::function<void(float)> listener);
     void writePcm(const char* buffer, long size);
 
     void deInitialize();
