+++ app/src/main/java/test/gleb_10_14_android/cppiface/SoundEnergyListener.java
@@ -1,5 +1,5 @@
 package test.gleb_10_14_android.cppiface;
 
 public interface SoundEnergyListener {
-    void onEnergyLevelCalculated(long level);
+    void onEnergyLevelCalculated(float level);
 }
